+++ ios/Classes/FlutterBraintreePlugin.h
@@ -1,4 +1,0 @@
-#import <Flutter/Flutter.h>
-
-@interface FlutterBraintreePlugin : NSObject<FlutterPlugin>
-@end