+++ ios/braintree_flutter_plus/Sources/braintree_flutter_plus/FlutterBraintreePlugin.swift
@@ -1,0 +1,7 @@
+import Flutter
+
+public class FlutterBraintreePlugin: NSObject, FlutterPlugin {
+    public static func register(with registrar: FlutterPluginRegistrar) {
+        FlutterBraintreeCustomPlugin.register(with: registrar)
+    }
+}